fix: aapt -k option not work under cygwin

+++ extract.sh
@@ -146,7 +146,13 @@ deodex() {
                 fi
             fi
         done
-        $aapt add -fk $deoappdir/$app/$app.apk $deoappdir/$app/$deoarch/classes*.dex || return 1
+        pushd $deoappdir/$app/$deoarch
+        adderror=false
+        $aapt add -fk ../../$app.apk classes*.dex || adderror=true
+        popd
+        if $adderror; then
+            return 1
+        fi
         apkfile=$deoappdir/$app/$app.apk
         $zipalign -f 4 $apkfile $apkfile-2 >/dev/null 2>&1
         mv $apkfile-2 $apkfile
@@ -184,7 +190,13 @@ deodex() {
             return 1
         fi
         $sevenzip d "$apkfile" $dexclass >/dev/null
-        $aapt add -fk $deoappdir/$app/$app.apk $deoappdir/$app/classes*.dex || return 1
+        pushd $deoappdir/$app
+        adderror=false
+        $aapt add -fk $app.apk classes*.dex || adderror=true
+        popd
+        if $adderror; then
+            return 1
+        fi
         rm -f $deoappdir/$app/classes.dex
         $zipalign -f 4 $apkfile $apkfile-2 >/dev/null 2>&1
         mv $apkfile-2 $apkfile
